requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.3.0
bleach==4.1.0
cffi==1.15.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-cors-headers==3.11.0
django-htmx==1.8.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gcloud==0.17.0
googleapis-common-protos==1.54.0
gunicorn==20.1.0
httplib2==0.20.4
itypes==1.2.0
Jinja2==3.0.3
jws==0.1.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.22.2
oauth2client==3.0.0
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.2
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.0
PyJWT==2.3.0
pyparsing==3.0.7
Pyrebase==3.0.27
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.11.1
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
rsa==4.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
webencodings==0.5.1
zipp==3.7.0
